requirements.txt
fastapi==0.115.12
ffmpeg==1.4
librosa==0.10.2.post1
numpy==2.2.6
Pillow==11.2.1
PyYAML==6.0.2
PyYAML==6.0.2
soundfile==0.12.1
torch==2.3.1+cu121
transformers==4.44.2
